% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Vehicle.R
\name{Bus}
\alias{Bus}
\title{R6 Child Class Representing a Bus}
\description{
Inherits from Vehicle. Vulnerable to standard road traffic.
}
\section{Super class}{
\code{WarsawTraffic::Vehicle} -> \code{Bus}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Bus-new}{\code{Bus$new()}}
\item \href{#method-Bus-check_disruption}{\code{Bus$check_disruption()}}
\item \href{#method-Bus-clone}{\code{Bus$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="WarsawTraffic" data-topic="Vehicle" data-id="update_location"><a href='../../WarsawTraffic/html/Vehicle.html#method-Vehicle-update_location'><code>WarsawTraffic::Vehicle$update_location()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bus-new"></a>}}
\if{latex}{\out{\hypertarget{method-Bus-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bus$new(id, line, lon, lat, last_update)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bus-check_disruption"></a>}}
\if{latex}{\out{\hypertarget{method-Bus-check_disruption}{}}}
\subsection{Method \code{check_disruption()}}{
Check if the bus route intersects with a traffic disruption
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bus$check_disruption(danger_polygon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{danger_polygon}}{An 'sf' polygon representing the blocked area}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Bus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
