% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_disruption_buffer}
\alias{create_disruption_buffer}
\title{Create a Spatial Buffer for a Disruption}
\usage{
create_disruption_buffer(lon, lat, radius_m = 15)
}
\arguments{
\item{lon}{A numeric value representing longitude (must be between -180 and 180).}

\item{lat}{A numeric value representing latitude (must be between -90 and 90).}

\item{radius_m}{A positive numeric value representing the buffer radius in meters. Default is 5.}
}
\value{
An \code{sfc} polygon object in EPSG:2180 projection representing the disruption zone.
}
\description{
Takes a raw GPS coordinate (longitude and latitude), converts it
into a spatial point, projects it to the metric Polish coordinate system (EPSG:2180),
and draws a circular buffer around it.
}
\examples{
\dontrun{
  # Create a n-meter buffer (15 meters default) around a point in central Warsaw
  danger_zone <- create_disruption_buffer(lon = 21.0122, lat = 52.2297, radius_m = 15)
}
}
