% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Vehicle.R
\name{Tram}
\alias{Tram}
\title{R6 Child Class Representing a Tram}
\description{
Inherits from Vehicle. Immune to road traffic. Flags as
blocked when its track is affected by a track blockage disruption.
}
\section{Super class}{
\code{\link[WarsawTraffic:Vehicle]{WarsawTraffic::Vehicle}} -> \code{Tram}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{is_blocked}}{Read-only. TRUE if tram is blocked by a track disruption.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tram-new}{\code{Tram$new()}}
\item \href{#method-Tram-check_disruption}{\code{Tram$check_disruption()}}
\item \href{#method-Tram-clear_delay}{\code{Tram$clear_delay()}}
\item \href{#method-Tram-clone}{\code{Tram$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="WarsawTraffic" data-topic="Vehicle" data-id="update_location"><a href='../../WarsawTraffic/html/Vehicle.html#method-Vehicle-update_location'><code>WarsawTraffic::Vehicle$update_location()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tram-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tram-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Tram object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tram$new(id, line, lon, lat, last_update)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tram-check_disruption"></a>}}
\if{latex}{\out{\hypertarget{method-Tram-check_disruption}{}}}
\subsection{Method \code{check_disruption()}}{
Check if this tram is affected by a disruption
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tram$check_disruption(affected_lines, disruption_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{affected_lines}}{A character vector of route_short_name values affected}

\item{\code{disruption_type}}{A string: "traffic" or "track_blockage"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tram-clear_delay"></a>}}
\if{latex}{\out{\hypertarget{method-Tram-clear_delay}{}}}
\subsection{Method \code{clear_delay()}}{
Reset this tram to non-delayed, non-blocked status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tram$clear_delay()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
