% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WarsawTraffic-package.R
\docType{package}
\name{WarsawTraffic-package}
\alias{WarsawTraffic}
\alias{WarsawTraffic-package}
\title{WarsawTraffic: Sumulator of the impact of traffic disruptions on the Warsaw public transit network.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Adam Jaworski \email{ap.jaworski3@student.uw.edu.pl}

}
\keyword{internal}
