% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_stale_data}
\alias{clean_stale_data}
\title{Clean stale vehicle data based on timestamp}
\usage{
clean_stale_data(df, max_age_mins = 10)
}
\arguments{
\item{df}{A data frame returned by fetch_warsaw_transit()}

\item{max_age_mins}{Maximum allowed age of the data point in minutes. Default is 10.}
}
\value{
A cleaned data frame with only recently updated vehicles.
}
\description{
Clean stale vehicle data based on timestamp
}
