% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fetcher.R
\name{fetch_warsaw_transit}
\alias{fetch_warsaw_transit}
\title{Fetch Live Public Transit Data from Warsaw API}
\usage{
fetch_warsaw_transit(api_key, vehicle_type = 1)
}
\arguments{
\item{api_key}{A string containing your personal API key from api.um.warszawa.pl}

\item{vehicle_type}{An integer: 1 for buses, 2 for trams. Default is 1.}
}
\value{
A data frame containing the live locations and details of the vehicles.
}
\description{
Fetch Live Public Transit Data from Warsaw API
}
