% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Vehicle.R
\name{Tram}
\alias{Tram}
\title{R6 Child Class Representing a Tram}
\description{
Inherits from Vehicle. Vulnerable only to track-specific disruptions.
}
\section{Super class}{
\code{WarsawTraffic::Vehicle} -> \code{Tram}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tram-new}{\code{Tram$new()}}
\item \href{#method-Tram-check_disruption}{\code{Tram$check_disruption()}}
\item \href{#method-Tram-clone}{\code{Tram$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="WarsawTraffic" data-topic="Vehicle" data-id="update_location"><a href='../../WarsawTraffic/html/Vehicle.html#method-Vehicle-update_location'><code>WarsawTraffic::Vehicle$update_location()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tram-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tram-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tram$new(id, line, lon, lat, last_update)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tram-check_disruption"></a>}}
\if{latex}{\out{\hypertarget{method-Tram-check_disruption}{}}}
\subsection{Method \code{check_disruption()}}{
Check if the tram is affected by a disruption
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tram$check_disruption(danger_polygon, disruption_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{danger_polygon}}{An 'sf' polygon representing the blocked area}

\item{\code{disruption_type}}{A character string (e.g., "traffic", "track_blockage")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
