% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Network.R
\name{TransitNetwork}
\alias{TransitNetwork}
\title{R6 Class Representing the Transit Network Manager}
\description{
Manages a fleet of Bus and Tram objects and handles global spatial updates.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransitNetwork-new}{\code{TransitNetwork$new()}}
\item \href{#method-TransitNetwork-update_fleet}{\code{TransitNetwork$update_fleet()}}
\item \href{#method-TransitNetwork-apply_disruption}{\code{TransitNetwork$apply_disruption()}}
\item \href{#method-TransitNetwork-clone}{\code{TransitNetwork$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitNetwork-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransitNetwork-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitNetwork$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitNetwork-update_fleet"></a>}}
\if{latex}{\out{\hypertarget{method-TransitNetwork-update_fleet}{}}}
\subsection{Method \code{update_fleet()}}{
Update the fleet with fresh API data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitNetwork$update_fleet(api_data, vechicle_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_data}}{A data frame returned by your API fetcher}

\item{\code{vechicle_type}}{An integer: 1 for buses, 2 for trams.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitNetwork-apply_disruption"></a>}}
\if{latex}{\out{\hypertarget{method-TransitNetwork-apply_disruption}{}}}
\subsection{Method \code{apply_disruption()}}{
Send a disruption warning to all vehicles in the fleet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitNetwork$apply_disruption(polygon, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{polygon}}{An 'sf' polygon representing the blocked area}

\item{\code{type}}{A string: "traffic" or "track_blockage"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitNetwork-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransitNetwork-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitNetwork$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
