% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Vehicle.R
\name{Vehicle}
\alias{Vehicle}
\title{R6 Parent Class Representing a Transit Vehicle}
\description{
A base class for buses and trams.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Vehicle-new}{\code{Vehicle$new()}}
\item \href{#method-Vehicle-update_location}{\code{Vehicle$update_location()}}
\item \href{#method-Vehicle-check_disruption}{\code{Vehicle$check_disruption()}}
\item \href{#method-Vehicle-clone}{\code{Vehicle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vehicle-new"></a>}}
\if{latex}{\out{\hypertarget{method-Vehicle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vehicle$new(id, line, lon, lat, last_update)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vehicle-update_location"></a>}}
\if{latex}{\out{\hypertarget{method-Vehicle-update_location}{}}}
\subsection{Method \code{update_location()}}{
Update the vehicle's GPS coordinates and timestamp
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vehicle$update_location(new_lon, new_lat, new_time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vehicle-check_disruption"></a>}}
\if{latex}{\out{\hypertarget{method-Vehicle-check_disruption}{}}}
\subsection{Method \code{check_disruption()}}{
Placeholder for the disruption logic to be overridden by child classes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vehicle$check_disruption(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vehicle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Vehicle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vehicle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
